requirements.txt
matplotlib==3.7.2
numpy==1.25.1
opencv-python-headless>=4.5.0
Pillow==9.4.0
streamlit==1.27.2
scipy==1.11.3
